Write request handlers that create admin user
The second page of the setup process was created. This page prompts
the user to enter the login credentials for the admin user. The request
handler for this page accepts POST requests containing admin user data,
creates a new admin user in the database, and then redirects the user
to the admin landing page.