Defines request handler for creating a new user
A new request handler was created that allows the admin to create new
user accounts. In the future, the admin will be able to assign different
roles to each user: some users with be admins, some will be authors, and
all others will be visitors. The User model was changed to accommodate
this. As a result, several older functions and methods had to be
adjusted to accommodate the new User model.

Several tests were written to test the request handler used to create
a new user. Also, other tests needed to be adjusted to account for the
new User model definition.