Remove files output by SASS and Webpack
The redundant CSS and JavaScript output files that are created by SASS
and Webpack were removed, as they are redundant. The script used to
initialize the app's configuration now uses SASS to create the app's
stylesheets from the project's SASS files. It also uses Webpack to
compile the front-end JavaScript files needed for the site to function
properly. This is all done automatically when the app is installed.