Modify admin site to allow for editing blog posts
The admin site should allow the admin to edit an existing blog
post and publish it. The GET and POST handlers for the admin page
that handles creating new posts were changed to allow for viewing
and updating an existing post. A new end-to-end test was written
to ensure this new functionality works as expected.