Allow admin to upload a main image for new posts
The admin should be allowed to upload a main image when creating
a new blog post. The npm package Multer was added to the list of
dependencies and configured to accept an image when a new blog
post is created through the admin site. A new end-to-end test was
written to ensure this new functionality works.