Define default settings before setup
Default settings were defined to be used in case a settings file
hasn't been created. This ensures the app still works even if the
npm run build command hasn't been run. One of the end-to-end tests
was dependent on a settings file being present in order to pass.
This test now passes regardless of whether a settings file exists,
ensuring that all tests are portable and able to run regardless
of the environment.