Change app to use new settings object.
The app now uses a settings object contructed using the new
Settings class. This Settings class defines an updateDatabase
instance method that actually returns a promise. The old
implementation did not return a promise and was prone to
errors. It allowed the server to make an asynchronous request to
the computer to update the contents of the settings file, but
would return a 200 reponse to the client without confirming that
the settings file was updated successfully. The new implementation
waits until the settings are updated before a response is sent.