Define new databaseUri getter method for Settings
The Settings class has a new getter method called databaseUri
that returns a string containing the URI for connecting to the
site's database. By doing this, it ensures that any function
can access the database URI from the Settings object. This
simplifies a lot of future code that needs to be written.