Create stylesheet for admin login page
A new SASS file was created that defines the styles for the admin
login page.

Furthermore, the build command was supposed to compile all SASS
files into the CSS files needed to style the site. The script
that performed this build task was not written correctly and so
several SASS files were not included during compilation. This has
been rectified and all SASS files are now compiled properly when
the build command is run.